+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 0.16.7
 
+* `Common`:
+    * New extensions `ifTrue`/`ifFalse`/`alsoIfTrue`/`alsoIfFalse`/`letIfTrue`/`letIfFalse`
+    * `Diff` now is serializable
+    * Add `IndexedValue` serializer
+    * `repeatOnFailure` extending: now you may pass any lambda to check if continue to try/do something
+* `Coroutines`:
+    * `Compose`:
+        * `asMutableComposeState` and all depending functions now use `doInUI` to guarantee state changin in Main Dispatcher
+
 ## 0.16.6
 
 * `Startup`: